+++ crate_universe/src/cli/generate.rs
@@ -13,6 +13,7 @@ use crate::lockfile::{lock_context, write_lockfile};
 use crate::metadata::{load_metadata, Annotations, Cargo};
 use crate::rendering::{write_outputs, Renderer};
 use crate::splicing::SplicingManifest;
+use crate::utils::normalize_cargo_file_paths;
 
 /// Command line options for the `generate` subcommand
 #[derive(Parser, Debug)]
@@ -77,8 +78,11 @@ pub fn generate(opt: GenerateOptions) -> Result<()> {
             let outputs = Renderer::new(config.rendering, config.supported_platform_triples)
                 .render(&context)?;
 
+            // make file paths compatible with bazel labels
+            let normalized_outputs = normalize_cargo_file_paths(outputs, &opt.repository_dir);
+
             // Write the outputs to disk
-            write_outputs(outputs, &opt.repository_dir, opt.dry_run)?;
+            write_outputs(normalized_outputs, opt.dry_run)?;
 
             return Ok(());
         }
@@ -116,8 +120,11 @@ pub fn generate(opt: GenerateOptions) -> Result<()> {
     )
     .render(&context)?;
 
-    // Write outputs
-    write_outputs(outputs, &opt.repository_dir, opt.dry_run)?;
+    // make file paths compatible with bazel labels
+    let normalized_outputs = normalize_cargo_file_paths(outputs, &opt.repository_dir);
+
+    // Write the outputs to disk
+    write_outputs(normalized_outputs, opt.dry_run)?;
 
     // Ensure Bazel lockfiles are written to disk so future generations can be short-circuited.
     if let Some(lockfile) = opt.lockfile {
