+++ crate_universe/src/utils.rs
@@ -6,7 +6,7 @@ pub(crate) mod target_triple;
 pub(crate) const CRATES_IO_INDEX_URL: &str = "https://github.com/rust-lang/crates.io-index";
 
 use std::collections::{BTreeMap, BTreeSet};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 /// Convert a string into a valid crate module name by applying transforms to invalid characters
 pub(crate) fn sanitize_module_name(name: &str) -> String {
@@ -31,6 +31,38 @@ pub(crate) fn sanitize_vendor_file_names(outputs: &BTreeMap<PathBuf, String>) ->
         .map(|p| {
             let p_str = sanitize_repository_name(p.to_str().unwrap());
             PathBuf::from(p_str)
+        })
+        .collect()
+}
+
+/// Vendored crates are generated by cargo itself in `src/metadata.rs` in the
+/// `VendorGenerator::generate()` method.  This means that the semver metadata will
+/// always contain a (+) symbol, which is not compatible with bazel's labels.  
+/// This function will rename the cargo vendor generated file paths to be compatible with bazel
+/// labels by simply replacing the (+) with a (-). If this file is called by any other cli mod,
+/// it just simplies joins the out dir to the path
+pub(crate) fn normalize_cargo_file_paths(
+    outputs: BTreeMap<PathBuf, String>,
+    out_dir: &Path,
+) -> BTreeMap<PathBuf, String> {
+    outputs
+        .into_iter()
+        .map(|(path, content)| {
+            let path = out_dir.join(path);
+            let original_path_str = path
+                .parent()
+                .expect("All file paths should have valid directories")
+                .to_str()
+                .expect("All file paths should be strings");
+            let new_path = if original_path_str.contains('+') {
+                let new_file_path = sanitize_repository_name(original_path_str);
+                std::fs::rename(original_path_str, &new_file_path).expect("Could not rename file");
+                PathBuf::from(new_file_path)
+            } else {
+                path
+            };
+
+            (new_path, content)
         })
         .collect()
 }
